split up large class file, remove service worker(??)
