+++ frontend/src/Map/index.jsx
@@ -7,8 +7,9 @@ import {
 } from 'react-leaflet'
 import { useState } from 'react'
 import { useMapEvent } from 'react-leaflet/hooks'
-import { domain } from '../actions.js'
-import { Intersection, Corridor } from '../spatialData.js'
+import { domain } from '../domain.js'
+import { Intersection } from '../intersection.js'
+import { Corridor } from '../corridor.js'
 import 'leaflet/dist/leaflet.css'
 
 const initialMapCenter = { lat: 43.65344, lng: -79.38400 }
