+++ frontend/src/index.jsx
@@ -1,12 +1,6 @@
 import * as ReactDOMclient from 'react-dom/client'
-import APP from './Layout'
-import * as serviceWorker from './serviceWorker'
+import App from './Layout'
 
 const container = document.getElementById('root')
 const root = ReactDOMclient.createRoot(container)
 root.render(<App/>)
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister()
